+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Import All has been added to public-facing API (#891)
+- Web UI workspace view now has an option to abort merge in progress (#895)
 
 ## Fixed
 - Web UI workspace view labels changes as Unmerged if there are merge conflicts (#890)